# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Kitna - Money clarity dashboard for creators and influencer managers
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Sign up a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me/settings:
    patch:
      operationId: updateSettings
      tags: [auth]
      summary: Update user settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsBody"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /deals:
    get:
      operationId: getDeals
      tags: [deals]
      summary: List all deals
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: income_type
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of deals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deal"

    post:
      operationId: createDeal
      tags: [deals]
      summary: Create a new deal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDealBody"
      responses:
        "201":
          description: Deal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /deals/{id}:
    get:
      operationId: getDeal
      tags: [deals]
      summary: Get a deal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deal details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DealWithPayments"
        "404":
          description: Deal not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateDeal
      tags: [deals]
      summary: Update a deal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDealBody"
      responses:
        "200":
          description: Deal updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
        "404":
          description: Deal not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteDeal
      tags: [deals]
      summary: Delete a deal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deal deleted
        "404":
          description: Deal not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        currency:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - currency
        - createdAt
        - updatedAt
    SignupBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
        currency:
          type: string
      required:
        - name
        - email
        - password
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    UpdateSettingsBody:
      type: object
      properties:
        name:
          type: string
        currency:
          type: string
        role:
          type: string
    Deal:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        brandName:
          type: string
        campaignName:
          type: string
        incomeType:
          type: string
        amountAgreed:
          type: string
        currency:
          type: string
        status:
          type: string
        isBarter:
          type: boolean
        barterEstimatedValue:
          type: ["string", "null"]
        barterDescription:
          type: ["string", "null"]
        taxableFlag:
          type: boolean
        startDate:
          type: ["string", "null"]
          format: date-time
        dueDate:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
        invoiceNumber:
          type: ["string", "null"]
        invoiceDate:
          type: ["string", "null"]
          format: date-time
        invoiceStatus:
          type: string
        invoiceAmount:
          type: ["string", "null"]
        managerCutType:
          type: ["string", "null"]
        managerCutValue:
          type: ["string", "null"]
        agencyCutType:
          type: ["string", "null"]
        agencyCutValue:
          type: ["string", "null"]
        gstApplicable:
          type: boolean
        taxableValue:
          type: ["string", "null"]
        tdsApplicable:
          type: boolean
        tdsAmount:
          type: ["string", "null"]
        totalReceived:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - brandName
        - campaignName
        - incomeType
        - amountAgreed
        - currency
        - status
        - isBarter
        - taxableFlag
        - invoiceStatus
        - gstApplicable
        - tdsApplicable
        - totalReceived
        - createdAt
        - updatedAt
    DealWithPayments:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        brandName:
          type: string
        campaignName:
          type: string
        incomeType:
          type: string
        amountAgreed:
          type: string
        currency:
          type: string
        status:
          type: string
        isBarter:
          type: boolean
        barterEstimatedValue:
          type: ["string", "null"]
        barterDescription:
          type: ["string", "null"]
        taxableFlag:
          type: boolean
        startDate:
          type: ["string", "null"]
          format: date-time
        dueDate:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
        invoiceNumber:
          type: ["string", "null"]
        invoiceDate:
          type: ["string", "null"]
          format: date-time
        invoiceStatus:
          type: string
        invoiceAmount:
          type: ["string", "null"]
        managerCutType:
          type: ["string", "null"]
        managerCutValue:
          type: ["string", "null"]
        agencyCutType:
          type: ["string", "null"]
        agencyCutValue:
          type: ["string", "null"]
        gstApplicable:
          type: boolean
        taxableValue:
          type: ["string", "null"]
        tdsApplicable:
          type: boolean
        tdsAmount:
          type: ["string", "null"]
        totalReceived:
          type: string
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - brandName
        - campaignName
        - incomeType
        - amountAgreed
        - currency
        - status
        - isBarter
        - taxableFlag
        - invoiceStatus
        - gstApplicable
        - tdsApplicable
        - totalReceived
        - payments
        - createdAt
        - updatedAt
    CreateDealBody:
      type: object
      properties:
        brandName:
          type: string
        campaignName:
          type: string
        incomeType:
          type: string
        amountAgreed:
          type: string
        currency:
          type: string
        isBarter:
          type: boolean
        barterEstimatedValue:
          type: ["string", "null"]
        barterDescription:
          type: ["string", "null"]
        taxableFlag:
          type: boolean
        startDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        invoiceNumber:
          type: ["string", "null"]
        invoiceDate:
          type: ["string", "null"]
        invoiceStatus:
          type: string
        invoiceAmount:
          type: ["string", "null"]
        managerCutType:
          type: ["string", "null"]
        managerCutValue:
          type: ["string", "null"]
        agencyCutType:
          type: ["string", "null"]
        agencyCutValue:
          type: ["string", "null"]
        gstApplicable:
          type: boolean
        taxableValue:
          type: ["string", "null"]
        tdsApplicable:
          type: boolean
        tdsAmount:
          type: ["string", "null"]
      required:
        - brandName
        - campaignName
        - incomeType
        - amountAgreed
    UpdateDealBody:
      type: object
      properties:
        brandName:
          type: string
        campaignName:
          type: string
        incomeType:
          type: string
        amountAgreed:
          type: string
        currency:
          type: string
        status:
          type: string
        isBarter:
          type: boolean
        barterEstimatedValue:
          type: ["string", "null"]
        barterDescription:
          type: ["string", "null"]
        taxableFlag:
          type: boolean
        startDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        invoiceNumber:
          type: ["string", "null"]
        invoiceDate:
          type: ["string", "null"]
        invoiceStatus:
          type: string
        invoiceAmount:
          type: ["string", "null"]
        managerCutType:
          type: ["string", "null"]
        managerCutValue:
          type: ["string", "null"]
        agencyCutType:
          type: ["string", "null"]
        agencyCutValue:
          type: ["string", "null"]
        gstApplicable:
          type: boolean
        taxableValue:
          type: ["string", "null"]
        tdsApplicable:
          type: boolean
        tdsAmount:
          type: ["string", "null"]
    Payment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        dealId:
          type: ["integer", "null"]
        amountReceived:
          type: string
        currency:
          type: string
        paymentMethod:
          type: string
        receivedDate:
          type: string
          format: date-time
        notes:
          type: ["string", "null"]
        tdsApplicable:
          type: boolean
        tdsAmount:
          type: ["string", "null"]
        dealBrandName:
          type: ["string", "null"]
        dealCampaignName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - amountReceived
        - currency
        - paymentMethod
        - receivedDate
        - tdsApplicable
        - createdAt
        - updatedAt
...