# lib/api-zod/src/generated/types/updateDealBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kitna - Money clarity dashboard for creators and influencer managers
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateDealBody {
  brandName?: string;
  campaignName?: string;
  incomeType?: string;
  amountAgreed?: string;
  currency?: string;
  status?: string;
  isBarter?: boolean;
  /** @nullable */
  barterEstimatedValue?: string | null;
  /** @nullable */
  barterDescription?: string | null;
  taxableFlag?: boolean;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  invoiceNumber?: string | null;
  /** @nullable */
  invoiceDate?: string | null;
  invoiceStatus?: string;
  /** @nullable */
  invoiceAmount?: string | null;
  /** @nullable */
  managerCutType?: string | null;
  /** @nullable */
  managerCutValue?: string | null;
  /** @nullable */
  agencyCutType?: string | null;
  /** @nullable */
  agencyCutValue?: string | null;
  gstApplicable?: boolean;
  /** @nullable */
  taxableValue?: string | null;
  tdsApplicable?: boolean;
  /** @nullable */
  tdsAmount?: string | null;
}